requirements.txt
# Vox Requirements

# PyObjC for macOS integration
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-AppKit>=10.0

# OpenAI API
openai>=1.0.0

# Configuration
pyyaml>=6.0
keyring>=24.0

# Build tools (for development)
py2app>=0.28

# Testing (for development)
pytest>=7.0
pytest-cov>=4.0

# Linting (for development)
ruff>=0.1.0
mypy>=1.0.0
